+++ src/pages/Detail.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import useInput from '../hooks/useInput';
-import { Route, Routes, useNavigate, useParams } from 'react-router-dom';
+import { useNavigate, useParams } from 'react-router-dom';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
 import { auth, db } from '../firebase';
 import { onAuthStateChanged } from 'firebase/auth';
@@ -8,19 +8,16 @@ import { collection, getDocs, query } from 'firebase/firestore';
 import { nanoid } from 'nanoid';
 import { deletePosts, getPosts } from '../api/posts';
 import { getComments, addComment } from '../api/comments';
-import { StCategory, StImage, StMpCategory, StTitle, StyledPostsyBox } from './Main';
 import { styled } from 'styled-components';
 import { StButton } from '../components/common/Button';
 import { StInput } from '../components/common/InputStyle';
 import { touristAttraction } from '../api/touristAttraction';
-import axios from 'axios';
-import { StCommentLi } from '../components/comment/commentStyle';
 import LoadingSpinner from '../components/common/LoadingSpinner';
 import Layout from '../components/common/Layout';
 import ReadMapInPost from '../components/Map/ReadMapInPost';
-import { useRecoilValue } from 'recoil';
-import { EmailAtom } from '../recoil/SearchAtom';
-import LikesPosts from '../components/Detail/LikesPosts';
+import DetailMain from '../components/Detail/DetailMain';
+import DetailOtherArea from '../components/Detail/DetailOtherArea';
+import DetailRecommendArea from '../components/Detail/DetailRecommendArea';
 
 const Detail = () => {
   const navigate = useNavigate();
@@ -38,20 +35,6 @@ const Detail = () => {
       queryClient.invalidateQueries('postsData');
     },
   });
-  // 더보기 관련 구간입니다.
-  const emailDisCharge = useRecoilValue(EmailAtom);
-  const handleSeeMoreButtonClick = () => {
-    setSeeMore(prev => !prev);
-  };
-
-  const handleUpdateButtonClick = async id => {
-    navigate(`/update/${id}`);
-  };
-
-  const handleDeleteButtonClick = async id => {
-    postsMutation.mutate(id);
-    navigate('/');
-  };
 
   // 코멘트 관련 입니다
   const [userEmail, setUserEmail] = useState('');
@@ -104,7 +87,7 @@ const Detail = () => {
       reset();
     }
   };
-  // 코멘드 관련 입니다
+  // 새로고침 관련 입니다
 
   useEffect(() => {
     if (!data) {
@@ -116,7 +99,7 @@ const Detail = () => {
   }, [isLoading, isError]);
 
   if (isLoading) {
-    return <div>로딩중 입니다.</div>;
+    return <LoadingSpinner />;
   }
 
   const posts = data.data.find(item => item.id == id);
@@ -134,49 +117,14 @@ const Detail = () => {
     }
   };
 
-  // 다른 지역 구간 입니다.
-  const handleDetailButtonClick = id => {
-    navigate(`/detail/${id}`);
-  };
-
   return (
     <Layout>
       {/* 메인 부분 */}
       <LoadingSpinner />
-      <StDetailMain>
-        <div>
-          <StDetailImage src={posts.image} />
-        </div>
-        <div>
-          <LikesPosts postId={posts.id} />
-          <StTitleDetail>
-            <StTitle>{posts.location}</StTitle>
-            {emailDisCharge == posts.email ? (
-              <div>
-                <StDetailSeeMore onClick={handleSeeMoreButtonClick} />
-                {seeMore && (
-                  <StDetaiSeeMorelUl>
-                    <StDetailListItem onClick={() => handleUpdateButtonClick(posts.id)}>
-                      <StDetailButton>수정</StDetailButton>
-                    </StDetailListItem>
-                    <StDetailListItem onClick={() => handleDeleteButtonClick(posts.id)}>
-                      <StDetailButton>삭제</StDetailButton>
-                    </StDetailListItem>
-                  </StDetaiSeeMorelUl>
-                )}
-              </div>
-            ) : null}
-          </StTitleDetail>
-          <div>
-            <StDetailCategory>지역: {posts.category}</StDetailCategory>
-            <p>{posts.description}</p>
-          </div>
-        </div>
-      </StDetailMain>
+      <DetailMain posts={posts} seeMore={seeMore} setSeeMore={setSeeMore} postsMutation={postsMutation} />
       {/* 메인 부분 */}
       {/* 추천 행사 구간 */}
       <ReadMapInPost posts={posts} />
-      <h1>gg</h1>
       {/* 코멘트 섹션입니다 */}
       <section>
         <div>
@@ -209,37 +157,10 @@ const Detail = () => {
       {/* 코멘트 섹션입니다 */}
 
       {/* 명소 지역 추천 입니다 . */}
-      <div>
-        <StRecommendTitle>명소 지역</StRecommendTitle>
-        <StDetailUl>
-          {list
-            ?.filter(item => item.firstimage !== '')
-            .slice(0, 5)
-            .map((item, index) => (
-              <StyledPostsyBox key={index}>
-                <StImage src={item.firstimage} />
-                <StTitle>{item.title}</StTitle>
-                <StMpCategory>{item.addr1}</StMpCategory>
-              </StyledPostsyBox>
-            ))}
-        </StDetailUl>
-      </div>
+      <DetailRecommendArea list={list} />
 
       {/* 다른 지역 구간입니다. */}
-      <div>
-        <StRecommendTitle>{posts.category} 다른 지역</StRecommendTitle>
-        <StDetailUl>
-          {data?.data
-            .filter(item => item !== posts && item.category === posts.category)
-            .map((item, index) => (
-              <StyledPostsyBox key={index} onClick={() => handleDetailButtonClick(item.id)}>
-                <StImage src={item.image} />
-                <StTitle>{item.location}</StTitle>
-                <StMpCategory>{item.category}</StMpCategory>
-              </StyledPostsyBox>
-            ))}
-        </StDetailUl>
-      </div>
+      <DetailOtherArea posts={posts} data={data} />
     </Layout>
   );
 };
@@ -250,11 +171,6 @@ const StCommentForm = styled.form`
   display: flex;
   align-items: baseline;
   gap: 10px;
-`;
-
-const StDetailMain = styled.main`
-  display: flex;
-  justify-content: space-between;
 `;
 
 const StComment = styled.li`
@@ -281,64 +197,6 @@ const StCommentRightPart = styled.div`
 
 const StCommentsContainer = styled.ol`
   margin-top: 40px;
-`;
-
-const StDetailCategory = styled.p`
-  color: black;
-  font-size: 16px;
-  font-weight: 800;
-  margin: 10px 0px;
-`;
-
-const StDetailImage = styled.img`
-  width: 190px;
-  height: 206px;
-  margin-bottom: 5px;
-  margin-right: 20px;
-`;
-
-const StTitleDetail = styled.div`
-  display: flex;
-  justify-content: space-between;
-`;
-
-const StDetaiSeeMorelUl = styled.ul`
-  position: fixed;
-`;
-
-const StDetailListItem = styled.li`
-  height: 40px;
-  padding: 5px 8px;
-  box-sizing: border-box;
-`;
-
-const StDetailButton = styled.button`
-  width: 100%;
-  padding: 7px 10px;
-  border: none;
-  background-color: #fff;
-  border-radius: 8px;
-  cursor: pointer;
-  text-align: left;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  overflow: hidden;
-
-  &:hover,
-  &:focus {
-    background-color: #f8e4ff;
-  }
-`;
-
-const StDetailSeeMore = styled.button`
-  width: 20px;
-  height: 20px;
-  border: none;
-  background-image: url(https://t1.daumcdn.net/tistory_admin/static/mobile/tt_img_area_reply.svg);
-  overflow: hidden;
-  font-size: 0;
-  line-height: 0;
-  color: transparent;
 `;
 
 export const StRecommendTitle = styled.p`
