+++ src/pages/Main.jsx
@@ -3,7 +3,6 @@ import { useQuery } from 'react-query';
 import { getPosts } from '../api/posts';
 import { styled } from 'styled-components';
 import { useNavigate } from 'react-router-dom';
-import { StButton } from '../components/common/Button';
 import { StyledMoodSelect } from './Write';
 import useInput from '../hooks/useInput';
 import LoadingSpinner from '../components/common/LoadingSpinner';
@@ -12,6 +11,8 @@ import Layout from '../components/common/Layout';
 import { useRecoilValue } from 'recoil';
 import { SearchAtom } from '../recoil/SearchAtom';
 import { getLikes } from '../api/likes';
+import Select from '../components/Main/Select';
+import PostBox from '../components/Main/PostBox';
 
 const Main = () => {
   const { isLoading, isError, data: posts } = useQuery('postsData', getPosts);
@@ -20,9 +21,6 @@ const Main = () => {
   const [postList, setPostList] = useState([]);
   const [likeList, setLikeList] = useState([]);
   const navigate = useNavigate();
-
-  // 전역 변수 사용
-  const searchList = useRecoilValue(SearchAtom);
 
   useEffect(() => {
     const post = posts?.data;
@@ -115,41 +113,8 @@ const Main = () => {
 
         <div>
           {/* 셀렉트박스  */}
-          <StyledMoodSelect onChange={setCategory}>
-            <option value="">전체</option>
-            <option value="서울">서울</option>
-            <option value="대구">대구</option>
-            <option value="부산">부산</option>
-            <option value="경기도">경기도</option>
-            <option value="강원도">강원도</option>
-            <option value="충청도">충청도</option>
-            <option value="경상도">경상도</option>
-            <option value="전라도">전라도</option>
-            <option value="제주도">제주도</option>
-          </StyledMoodSelect>
-          <StUl>
-            {searchList.length > 0
-              ? searchList.map((item, index) => (
-                  <StyledPostsyBox key={index} onClick={() => handleDetailButtonClick(item.id)}>
-                    <div>
-                      <StImage src={item.image} />
-                      <StTitle>{item.location}</StTitle>
-                      <StMpCategory>{item.category}</StMpCategory>
-                    </div>
-                  </StyledPostsyBox>
-                ))
-              : posts.data
-                  ?.filter(item => !category || item.category === category)
-                  .map((item, index) => (
-                    <StyledPostsyBox key={index} onClick={() => handleDetailButtonClick(item.id)}>
-                      <div>
-                        <StImage src={item.image} />
-                        <StTitle>{item.location}</StTitle>
-                        <StMpCategory>{item.category}</StMpCategory>
-                      </div>
-                    </StyledPostsyBox>
-                  ))}
-          </StUl>
+          <Select setCategory={setCategory} />
+          <PostBox posts={posts} category={category} handleDetailButtonClick={handleDetailButtonClick} />
         </div>
       </Layout>
     </>
