+++ src/pages/MapTest.jsx
@@ -1,6 +1,4 @@
 import React from 'react';
-import ReadMapInPost from '../components/Map/ReadMapInPost';
-import TestMap from '../components/Map/TestMap';
 import MapForWrite from '../components/Map/MapForWrite';
 import Layout from '../components/common/Layout';
