+++ src/pages/UpdateWrite.jsx
@@ -1,15 +1,11 @@
 import React, { useEffect, useState } from 'react';
-import { StSearchText } from '../components/common/Header';
-import { styled } from 'styled-components';
 import useInput from '../hooks/useInput';
-import { StButton } from '../components/common/Button';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
 import { getPosts, updatePosts } from '../api/posts';
 import { useNavigate, useParams } from 'react-router-dom';
-import MapForUpdate from '../components/Map/MapForUpdate';
-import { StRequiredFieldsText } from './Write';
 import LoadingSpinner from '../components/common/LoadingSpinner';
 import Layout from '../components/common/Layout';
+import UpdateForm from '../components/UpdateWrite/UpdateForm';
 
 function UpdateWrite() {
   const navigate = useNavigate();
@@ -45,6 +41,10 @@ function UpdateWrite() {
 
   if (isLoading) {
     return <div>로딩중 입니다..</div>;
+  }
+
+  if (isError) {
+    return <h1>에러가 발생하였습니다.</h1>;
   }
 
   const posts = data.data.find(item => item.id == id);
@@ -85,58 +85,29 @@ function UpdateWrite() {
     postsMutation.mutate({ id, posts, newPosts });
     navigate('/');
   };
-  const handleCancelButtonClick = () => {
-    navigate(`/detail/${id}`);
-  };
-
-  if (isError) {
-    return <h1>에러가 발생하였습니다.</h1>;
-  }
 
   return (
     <>
       <LoadingSpinner />
 
       <Layout>
-        <form onSubmit={handleWriteButtonClick}>
-          지역:
-          <StyledMoodSelect onChange={setCategory}>
-            <option value="서울">서울</option>
-            <option value="대구">대구</option>
-            <option value="부산">부산</option>
-            <option value="경기도">경기도</option>
-            <option value="강원도">강원도</option>
-            <option value="충청도">충청도</option>
-            <option value="경상도">경상도</option>
-            <option value="전라도">전라도</option>
-            <option value="제주도">제주도</option>
-          </StyledMoodSelect>
-          <div>
-            location <br />
-            <StRequiredFieldsText value={updatelocation} onChange={setUpdateLocation} />
-          </div>
-          <div>
-            description <br />
-            <StRequiredFieldsText value={updatedescription} onChange={setUpdateDescription} />
-          </div>
-          <div>
-            Image <br />
-            <StRequiredFieldsText value={updateimage} onChange={setUpdateImage} />
-          </div>
-          <MapForUpdate markerInfo={markerInfo} setMarkerInfo={setMarkerInfo} posts={posts} />
-          <StButton $fontColor={'black'}>수정</StButton>
-          <StButton type="button" onClick={handleCancelButtonClick} $fontColor={'black'}>
-            취소{' '}
-          </StButton>
-        </form>
+        <UpdateForm
+          handleWriteButtonClick={handleWriteButtonClick}
+          setCategory={setCategory}
+          updatelocation={updatelocation}
+          setUpdateLocation={setUpdateLocation}
+          updatedescription={updatedescription}
+          setUpdateDescription={setUpdateDescription}
+          updateimage={updateimage}
+          setUpdateImage={setUpdateImage}
+          markerInfo={markerInfo}
+          setMarkerInfo={setMarkerInfo}
+          posts={posts}
+          id={id}
+        />
       </Layout>
     </>
   );
 }
 
 export default UpdateWrite;
-
-const StyledMoodSelect = styled.select`
-  padding: 10px;
-  font-size: 16px;
-`;
