+++ src/pages/Write.jsx
@@ -1,16 +1,14 @@
 import React, { useEffect, useState } from 'react';
-import { StSearchText } from '../components/common/Header';
 import { styled } from 'styled-components';
 import useInput from '../hooks/useInput';
-import { StButton } from '../components/common/Button';
 import { useMutation, useQueryClient } from 'react-query';
 import { addPosts } from '../api/posts';
 import { useNavigate } from 'react-router-dom';
-import MapForWrite from '../components/Map/MapForWrite';
 import LoadingSpinner from '../components/common/LoadingSpinner';
 import Layout from '../components/common/Layout';
 import { useRecoilValue } from 'recoil';
 import { EmailAtom } from '../recoil/SearchAtom';
+import WriteForm from '../components/Write/WriteForm';
 
 function Write() {
   const navigate = useNavigate();
@@ -33,6 +31,7 @@ function Write() {
   useEffect(() => {
     console.log(emailDischarge);
   }, []);
+
   const handleWriteButtonClick = e => {
     e.preventDefault();
     if (!markerInfo) {
@@ -59,52 +58,21 @@ function Write() {
     <>
       <LoadingSpinner />
       <Layout>
-        <form onSubmit={handleWriteButtonClick}>
-          지역:
-          <StyledMoodSelect onChange={setCategory}>
-            <option value="서울">서울</option>
-            <option value="대구">대구</option>
-            <option value="부산">부산</option>
-            <option value="경기도">경기도</option>
-            <option value="강원도">강원도</option>
-            <option value="충청도">충청도</option>
-            <option value="경상도">경상도</option>
-            <option value="전라도">전라도</option>
-            <option value="제주도">제주도</option>
-          </StyledMoodSelect>
-          <div>
-            location <br />
-            <StRequiredFieldsText value={location} onChange={setLocation} />
-          </div>
-          <div>
-            description <br />
-            <StRequiredFieldsText value={description} onChange={setDescription} />
-          </div>
-          <div>
-            Image <br />
-            <StRequiredFieldsText value={image} onChange={setImage} />
-          </div>
-          <MapForWrite markerInfo={markerInfo} setMarkerInfo={setMarkerInfo} />
-          <StButton $fontColor={'black'}>등록</StButton>
-        </form>
+        <WriteForm
+          handleWriteButtonClick={handleWriteButtonClick}
+          setCategory={setCategory}
+          location={location}
+          setLocation={setLocation}
+          description={description}
+          setDescription={setDescription}
+          image={image}
+          setImage={setImage}
+          markerInfo={markerInfo}
+          setMarkerInfo={setMarkerInfo}
+        />
       </Layout>
     </>
   );
 }
 
 export default Write;
-
-export const StyledMoodSelect = styled.select`
-  padding: 10px;
-  font-size: 16px;
-`;
-export const StRequiredFieldsText = styled.input`
-  width: 800px;
-  height: 60px;
-  border: 1px solid #ccc;
-  border-radius: 12px;
-  margin-right: 10px;
-  margin-top: 10px;
-  font-size: 18px;
-  margin-bottom: 10px;
-`;
